+++ CHANGELOG.md
@@ -24,6 +24,8 @@ and this project adheres to
 * Added size limiting to `EthernetFrameClass` write functions.
 * New `EthernetClass::waitForLink(timeout)` function that waits for a link to
   be detected.
+* Added a way to allow or disallow receiving frames addressed to specific MAC
+  addresses: `EthernetClass::setMACAddressAllowed(mac, flag)`
 
 ### Changed
 * Updated `SNTPClient` example to use `EthernetUDP::send()` instead of